requirements.txt
paho-mqtt==1.5.0
rhasspy-hermes~=0.3.0
rhasspy-wake-raven~=0.1.0
